fix replace singular url in text
